break harness up into modules
