+++ components/lark-test/src/harness/test/execution_test.rs
@@ -1,0 +1,7 @@
+use crate::harness::test::TestContext;
+
+impl TestContext<'_> {
+    crate fn run_execute_test(&self) {
+        unimplemented!("execution test")
+    }
+}
