+++ components/lark-test/src/lib.rs
@@ -13,6 +13,7 @@ use std::fmt::Debug;
 
 mod harness;
 pub use harness::run_test_harness;
+pub use harness::search_files;
 
 pub use lark_debug_with::DebugWith;
 pub use lark_span::IntoFileName;
